requirements.txt
certifi==2018.8.24
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.4
django-filter==2.0.0
djangorestframework==3.8.2
ez-setup==0.9
gunicorn==19.7.1
idna==2.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
mysqlclient==1.3.12
Pillow==5.1.0
pkg-resources==0.0.0
pytz==2018.4
requests==2.19.1
uritemplate==3.0.0
urllib3==1.23
